% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoskosis.R
\name{k_get_date_list}
\alias{k_get_date_list}
\title{R Package for scraping data from ECOS and KOSIS
k_get_date_list}
\usage{
k_get_date_list(start_date, end_date)
}
\description{
R Package for scraping data from ECOS and KOSIS
k_get_date_list
}
\author{
Junah Bang \email{juna0033@g.skku.edu}
}
