% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoskosis.R
\name{kosis_get_api_db}
\alias{kosis_get_api_db}
\title{get api db from kosis; fot target period}
\usage{
kosis_get_api_db(api_key, start_date, end_date, stat_id_list)
}
\description{
get api db from kosis; fot target period
}
